Update non-MEG channel names and locations when importing BIDS files (#810)
* Add electrode positions from `_electrodes.tsv` after finishing `_channels.tsv` tasks

As explained in the header, if channel names are different in data and BIDS metadata then these locations won't be updated. So, it makes more sense to first finish parsing `_channels.tsv` (which contains the actual names of the channels), do the necessary channel name changes in the channel file (if any) and then import in the locations from `_electrodes.tsv` (which has the same channel name as in `_channels.tsv`).

* Update EEG channel names for `.fif` files from `_channels.tsv`

* For all non-MEG, make _channels.tsv authoritative